+++ projects/ngx-dmaster-ui/src/lib/components/primitives/chip/index.ts
@@ -1,0 +1,4 @@
+export * from './chip.component';
+export * from './chip-set.component';
+export * from './chip.tokens';
+export * from './chip.types';